requirements/base.txt
django==5.0.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.0
redis==5.0.0
Pillow==10.2.0
